[ ODS-5889][ODS-5919] Bring in EdOrgID size change & Fix smoke test issue (#770)
* Convert all usages of education organization ids in explicit (non-generated) code to longs.

* Eliminate ResourceHelper whose methods are all unused except CreateStudent, which is now folded into StudentTests class where it is used exclusively.

* Removed unused method in EducationOrganizationHelper.

* Added a check in Open API metadata tests for the formatting indicating that the SchoolId now has a "format" of int64.

* Added idempotent scripts for altering the EducationOrganizationId column in the Admin database to bigint for Postgres and SQL Server.

* Bring in EdOrgID size change - Standard Core Changes

* EdFi.Suite3.Admin.DataAccess and EdFi.Suite3.Ods.Common and EdFi.Suite3.Common package version updated

* Script updated

* 0150-Alter-EducationOrganizationId-to-bigint.sql file Moved to /5.0.0/Artifacts/MsSql/Structure/Admin

* EdFi.Suite3.Admin.DataAccess package version updated

* EdFi.Suite3.Common package version updated

* EdFi.Suite3.Admin.DataAccess package version updated

* dbo.GetClientForKey Function updated with EducationOrganizationId BIGINT

* Script updated

* dbo.GetClientForToken function updated with BigInt

* Long DataType Added

* Pkg EdFi.Ods.Populated.Template updated with Debug

* CONFIGURATION: "Debug" updated

* EducationOrganizationIdToEducationOrganizationId table updated with Bigint

* CONFIGURATION: "Release" updated

* auth.EducationOrganizationIdToEducationOrganizationId table updated with BigInt

* Added idempotent scripts for altering the EducationOrganizationId column in the Admin database to bigint for Postgres and SQL Server.

* EdFi.Suite3.Admin.DataAccess and EdFi.Suite3.Ods.Common and EdFi.Suite3.Common package version updated

* Script updated

* 0150-Alter-EducationOrganizationId-to-bigint.sql file Moved to /5.0.0/Artifacts/MsSql/Structure/Admin

* EdFi.Suite3.Admin.DataAccess package version updated

* EdFi.Suite3.Common package version updated

* EdFi.Suite3.Admin.DataAccess package version updated

* Long DataType Added

* Pkg EdFi.Ods.Populated.Template updated with Debug

* CONFIGURATION: "Debug" updated

* EducationOrganizationIdToEducationOrganizationId table updated with Bigint

* CONFIGURATION: "Release" updated

* Admin Alter EductionOrganizantionID to bigint moved Scripts

* EdFi.Suite3.Common package version updated

* EdFi.Suite3.Admin.DataAccess package version updated

* CSharpDefaultHasDomainMeaning added long check

* EdFi.Ods.Api.IntegrationTests project  Bigint Code fix for Both DataStandard

* EdFi.Ods.Repositories.NHibernate.Tests and EdFi.Ods.Tests projects codefix for BigInt

* Should_contain_a_definition_for_academicWeek_key codefix

* academicWeeks Empty SchoolID Validation Error Message added

* Should return the item with SchoolId of the resource is empty string message updated

* Updated ApprovalTest

* Empty-Commit

* FiscalYear 4.0.0 Message Codefix

* [ODS-5919] Fix smoke test issue

* FiscalYear 4.0.0 Message Codefix

---------

+++ Application/EdFi.Admin.DataAccess/Models/Application.cs
@@ -42,7 +42,7 @@ namespace EdFi.Admin.DataAccess.Models
 
         public virtual ICollection<Profile> Profiles { get; set; }
 
-        public ApplicationEducationOrganization CreateApplicationEducationOrganization(int educationOrganizationId)
+        public ApplicationEducationOrganization CreateApplicationEducationOrganization(long educationOrganizationId)
             => new ApplicationEducationOrganization
             {
                 EducationOrganizationId = educationOrganizationId,
