+++ Application/EdFi.Admin.DataAccess/Models/ApplicationEducationOrganization.cs
@@ -23,7 +23,7 @@ namespace EdFi.Admin.DataAccess.Models
 
         public Application Application { get; set; }
 
-        public int EducationOrganizationId { get; set; }
+        public long EducationOrganizationId { get; set; }
 
         public ICollection<ApiClient> Clients { get; set; }
     }
