+++ Application/EdFi.Ods.Api/Security/Authentication/RawApiClientDetailsDataRow.cs
@@ -17,7 +17,7 @@ namespace EdFi.Ods.Api.Security.Authentication
 
         public string StudentIdentificationSystemDescriptor { get; set; }
 
-        public int? EducationOrganizationId { get; set; }
+        public long? EducationOrganizationId { get; set; }
 
         public string ClaimSetName { get; set; }
